# schemas/control-plane-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Proxy Sentinel Control Plane API
  version: 1.0.0
  description: Control-plane contract for the Proxy Sentinel frontend. The API exposes normalized events, evidence, risk snapshots, labels, shadow runs, audit logs, and session state.
servers:
  - url: /api/v1
paths:
...
  /integrations/identity/snapshots:
    post:
      operationId: commitIdentitySnapshot
      description: Atomically submit one complete online inventory for a source, sensor, campus and access domain. Requires the identity integration token and respects the global read-only switch. Incomplete inventories never close sessions. Maximum 10000 sessions per scope; do not split an inventory into independently complete requests.
      security:
        - IdentityBearer: []
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentitySnapshotRequest' }
      responses:
        "202":
          description: Complete snapshot and audit record committed atomically.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdentitySnapshotReceipt' }
        "200":
          description: Identical snapshot was already committed; freshness is not advanced by retry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdentitySnapshotReceipt' }
        "400": { description: 'Incomplete inventory, count mismatch, invalid record, scope conflict or invalid observation time.' }
        "401": { description: Invalid identity integration token. }
        "403": { description: Global read-only mode. }
        "409": { description: 'Snapshot identity or observation time already contains different content, or storage does not support reconciliation.' }
        "503": { description: Database commit failed or its result is uncertain; retry the same immutable request and key. }
  /integrations/identity/snapshot-uploads:
    post:
      operationId: createIdentitySnapshotUpload
      description: Creates a bounded staged upload for a complete identity snapshot of up to 200000 address records.
      security:
        - IdentityBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentitySnapshotUploadCreate' }
      responses:
        '202': { description: Upload reservation created. }
        '200': { description: Identical upload reservation already exists. }
...
  /accounts/{account_id}/identity:
    get:
      operationId: getAccountIdentity
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/SensorIdQuery"
        - $ref: "#/components/parameters/WindowQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        "200":
          description: Current account identity profile with linked endpoints, IPs and access locations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountIdentityProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /endpoints/{endpoint_id}/identity:
    get:
      operationId: getEndpointIdentity
      parameters:
        - name: endpoint_id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/SensorIdQuery"
        - $ref: "#/components/parameters/WindowQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        "200":
          description: Current endpoint identity profile with linked accounts, IPs and access locations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndpointIdentityProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    IdentityBearer:
      type: http
      scheme: bearer
      description: Dedicated RADIUS/Portal integration credential; never a user session token.
    ProductPolicyBearer:
      type: http
      scheme: bearer
      description: Dedicated read-only product/control snapshot integration credential.
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
...
    SensorIdQuery:
      name: sensor_id
      in: query
      schema:
        type: string
...
    FromQuery:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    ToQuery:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    WindowQuery:
      name: window
      in: query
      schema:
        type: string
        enum: [10m, 1h, 24h]
...
  schemas:
...
    IdentitySnapshotRequest:
      type: object
      additionalProperties: false
      required: [source, sensor_id, campus_id, access_domain, observed_at, reconcile_interval_seconds, complete, expected_count, records]
      properties:
        source: { type: string, minLength: 1, maxLength: 200 }
        sensor_id: { type: string, minLength: 1, maxLength: 200 }
        campus_id: { type: string, minLength: 1, maxLength: 200 }
        access_domain: { type: string, minLength: 1, maxLength: 200 }
        observed_at:
          type: string
          format: date-time
          description: Consistent upstream inventory time, within the past seven days, UTC with at most microsecond precision; never a fetch-completion timestamp for an inconsistent scan.
        reconcile_interval_seconds: { type: integer, minimum: 1, maximum: 86400 }
        complete: { type: boolean, enum: [true] }
        expected_count: { type: integer, minimum: 0, maximum: 10000 }
        records:
          type: array
          maxItems: 10000
          items:
            type: object
            required: [session_id, account_id, ip]
            properties:
              session_id: { type: string, minLength: 1 }
              account_id: { type: string, minLength: 1 }
              ip: { type: string }
              endpoint_id: { type: string }
              mac: { type: string }
              group_id: { type: string }
              product_id: { type: string }
              device_class: { type: string, enum: [mobile, pc, other] }
            additionalProperties: { type: string }
    IdentitySnapshotUploadCreate:
      type: object
      additionalProperties: false
      required: [upload_id, source, sensor_id, observed_at, reconcile_interval_hours, expected_count, expected_chunks, expected_sha256]
      properties:
        upload_id: { type: string, minLength: 1, maxLength: 200 }
        source: { type: string, minLength: 1, maxLength: 200 }
        sensor_id: { type: string, minLength: 1, maxLength: 200 }
        campus_id: { type: string, maxLength: 200 }
        access_domain: { type: string, maxLength: 200 }
        observed_at: { type: string, format: date-time }
        reconcile_interval_hours: { type: integer, minimum: 1, maximum: 168 }
        expected_count: { type: integer, minimum: 0, maximum: 200000 }
        expected_chunks: { type: integer, minimum: 1, maximum: 200 }
        expected_sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    IdentitySnapshotReceipt:
      type: object
      required: [snapshot_id, status, observed_at, session_count, replayed]
      properties:
        snapshot_id: { type: string }
        status: { type: string, enum: [completed] }
        observed_at: { type: string, format: date-time }
        session_count: { type: integer }
        replayed: { type: boolean }
...
    DeviceName:
      type: object
      required: [value, source, manual, status, multiple_names]
      properties:
        value:
          type: string
        source:
          type: string
        manual:
          type: boolean
        status:
          type: string
          enum: [current, historical]
        multiple_names:
          type: boolean
...
    DiscoveryRecognitionSummary:
      type: object
      properties:
        capabilities: { type: array, items: { type: string } }
        types: { type: array, items: { type: string } }
        positions: { type: array, items: { type: string } }
        conflict: { type: boolean }
        evidence_count: { type: integer }
...
    EndpointEntity:
      type: object
      required: [endpoint_id, entity_role, first_seen, last_seen, identity_confidence]
      properties:
        endpoint_id:
          type: string
        primary_mac:
          type: string
        entity_role:
          type: string
          enum: [endpoint, infrastructure, gateway, nat, server, network_device, unknown]
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        identity_confidence:
          type: number
          minimum: 0
          maximum: 1
        attributes:
          type: object
          additionalProperties: true
        registration_status:
          type: string
          enum: [unregistered, registered, ignored, retired]
        owner_account:
          type: string
        owner_name:
          type: string
        owner_department:
          type: string
        asset_tag:
          type: string
        ownership_class:
          type: string
          enum: [unknown, byod, school_asset, public_terminal, infrastructure]
        registered_by:
          type: string
        registered_at:
          type: string
          format: date-time
        registration_note:
          type: string
        merge_status:
          type: string
          enum: [active, merged, split]
        merged_into_endpoint_id:
          type: string
        split_from_endpoint_id:
          type: string
        registration_updated_at:
          type: string
          format: date-time
        registration_update_by_id:
          type: string
...
    AccountSession:
      type: object
      required: [session_id, account_id, endpoint_id, entity_role, started_at, ended_at, identity_confidence, source]
      properties:
        session_id:
          type: string
        account_id:
          type: string
        endpoint_id:
          type: string
        entity_role:
          type: string
          enum: [endpoint, infrastructure, gateway, nat, server, network_device, unknown]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        identity_confidence:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
    IdentityIPMACHistory:
      type: object
      required: [event_id, endpoint_id, account_id, entity_role, seen_at, source]
      properties:
        event_id:
          type: string
        endpoint_id:
          type: string
        account_id:
          type: string
        ip:
          type: string
        mac:
          type: string
        entity_role:
          type: string
          enum: [endpoint, infrastructure, gateway, nat, server, network_device, unknown]
        seen_at:
          type: string
          format: date-time
        source:
          type: string
    IdentityAccessHistory:
      type: object
      required: [event_id, endpoint_id, account_id, access_id, seen_at, source]
      properties:
        event_id:
          type: string
        endpoint_id:
          type: string
        account_id:
          type: string
        access_id:
          type: string
        vlan_id:
          type: string
        switch_id:
          type: string
        port_id:
          type: string
        ap_id:
          type: string
        ssid:
          type: string
        seen_at:
          type: string
          format: date-time
        source:
          type: string
    AccountIdentityProfile:
      type: object
      required: [account_id, summary, endpoints, sessions, ip_history, access_history]
      properties:
        account_id:
          type: string
        summary:
          type: string
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/EndpointEntity"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/AccountSession"
        ip_history:
          type: array
          items:
            $ref: "#/components/schemas/IdentityIPMACHistory"
        access_history:
          type: array
          items:
            $ref: "#/components/schemas/IdentityAccessHistory"
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
    EndpointIdentityProfile:
      type: object
      required: [endpoint_id, summary, endpoint, accounts, sessions, ip_history, access_history]
      properties:
        discovery:
          $ref: '#/components/schemas/DiscoveryRecognitionSummary'
        device_names_disabled:
          type: boolean
        device_name:
          $ref: '#/components/schemas/DeviceName'
        recognition: { $ref: "#/components/schemas/EndpointDeviceInventory" }
        brand_inference: { $ref: "#/components/schemas/BrandInference" }
        endpoint_id:
          type: string
        summary:
          type: string
        endpoint:
          $ref: "#/components/schemas/EndpointEntity"
        accounts:
          type: array
          items:
            type: string
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/AccountSession"
        ip_history:
          type: array
          items:
            $ref: "#/components/schemas/IdentityIPMACHistory"
        access_history:
          type: array
          items:
            $ref: "#/components/schemas/IdentityAccessHistory"
        ecosystem_evidence:
          type: array
          items:
            $ref: "#/components/schemas/EndpointDomainEvidence"
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
    EndpointDomainEvidence:
      type: object
      required: [endpoint_id, domain, ecosystem, event_source, attribution_method, rule_source, rule_version, category, confidence, first_seen, last_seen, count]
      properties:
        rule_domain: { type: string }
        source_url: { type: string }
        source_version: { type: string }
        service: { type: string }
        purpose: { type: string }
        brand_eligible: { type: boolean }
        endpoint_id: { type: string }
        ip: { type: string }
        auth_session_id: { type: string }
        domain: { type: string }
        ecosystem: { type: string }
        event_source: { type: string }
        attribution_method: { type: string }
        rule_source: { type: string }
        rule_version: { type: string }
        category: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        count: { type: integer }
        event_ids_sample:
          type: array
          items: { type: string }
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string